Introduce SIMS worker
Wherein we implement a new worker binary called SIMS (Signal informed
Messaging Service) which consumes the state changes emitted by the API
and by applying rules determines who to send transactional messages to.
In this first iteration we only target mobile platforms with push
messages. As the delivery platform SNS is supported and used for the
bookkeeping of endpoint per device.

* implment SIMS binary
* add rules for:
  - new follower
  - new friend request
  - friend request confirmed
  - new like
  - new comment
  - new post
* introduce SNS as push delivery platform