Add infrastructure description
Wherein we introduce a terraform template which allows to spin up the
entire Tapglue stack on AWS. As this is the extraction of how the
infrastructure was operated for the Tapglue platform it comes with quite
some assumptions regarding isolation, security and network layout.
Therefore this template is best used in isolation on a sub-account to
not interfer with existing installations. To do there is a companion CLI
to help run the stack with a single command. *Currently not encoded are
the creations of certain key AMIs*.

* teraform template for Tapglue stack
* terraform command for setup, teardown & update