Simplify infrastructure
Wherein we try to simplify the infrastructure to the point where we
don't need custom AMIs and extensive tooling to create Docker
containers. We achieve the first one by using of-the-shelf AMIs offered
in marketplaces and/or advertised by AWS. The second one is solved with
using a public repository[0] on DockerHub.

* use pre-maded images
* use configured SSH key
* introduce simple TODO list

[0] https://hub.docker.com/r/tapglue/snaas/