Implemenet data migration tool
Wherein we introduce dataz a binary to export and import data, which in
the first iteration uses newline delimited json files per entity. It
supports export and import as commands which aim to work symmetrically.
We choose the application as capsule for export and import to avoid more
involved migration coordination/logic. The user must be aware of the
namespace of the application that they want to export, where on import a
new application is created and the namespace is logged.

Example usage:
./dataz \
  -data.dir "app_1_2" \
  -namespace "app_1_2" \
  -postgres.url "$PG_URL"
  export