Support remote state on setup
As many use remote state to enable multiple collaborators to work on the
same state we're introducing the support for it in terraformer. This
currently entails setting up a versioned bucket and run
`terraform remote config`, we attacked s3 first as it is natural with
the general focus on AWS. There are still a lot of things to iron out
and we likely would want to support user input on state configuration
concerns.