Reactions
To avoid a singular type of reaction to Posts which ultimately leads to
a gap in the expressiveness for users we want to follow the prominent
example of other social networks and make a range of emotions available
to our customers which they can expose as interactions on Posts and
potentially other entities like Comments and Events themselves in the
future. The proposed set of reactions includes six initial reactions
including the already supported likes.

* implement Reaction service
* implement Reaction handlers
* implement react on a Post
* enrich Posts with Reaction counts