Add Prometheus based monitoring
This sets up Prometheus which scrapes the service binaries and host data
which then is presented in an operation dashboard in Grafana. For easy
access we provide another public endpoint via:

monitoring-$env-$region-$domain

While we want simple availability it should not be exposed to the
greater public, therefore the Grafana installation is guarded with a
Google login. This forces us to require a Google OAuth client id and
secret during setup.

* setup prometheus
* setup grafana
* setup operations dashboard