Invite flow
To enable a better on-boarding experience we need a way to correlate
users and their inviter. One way is to keep a list of invites of a user
that stores a unique id (e.g. facebook id, or generate random identifier)
which can be kept through until the invitee hits the signup. If the
invitee signs up and the conenction to the inviter is preserved we can
automatically create connections for those so they can hit the ground
running.