Carry precompiled templates from notification pipeline
To finally support all the local notifications we need a way to send
the configured template per language in the actual notification. This
change compiles all templates for all messages before returning it as
part of a map from the pipeline. This could certainly be improved to not
do unnecessary work but would need that the channel which has the
device knowledge has to know about the templates, either way could be
wasteful. For now we trade cpu time for complexity.