Add Counters
This adds a simple way to have namespaced counters per user and get the
global count per namespace. In essence this can be used for a variety of
use-cases, e.g. leaderboards, group challenges, etc. Introducing only
the barest of functionality it will be extended as use-cases arise and
applications have been identified. A variety of dimensions might be
added for future extension to have contextual leadboards, for example
slicing by location, user core data or free form tags.

Counter payload:

```
{
  "value": 123
}
```

In the first iteration we support two endpoints:

* `PUT /me/counters/<counterName>`: Expects the counter payload and sets
the named counter for the current user to the given value.
* `GET /counters/<counterName>`: Returns the sum of all values for all
users for the named counter.