Implement reaction multi count
Since the introduction of reactions one of the biggest impacts on
response times for feed/list requests has been the constant overhead of
countless network requests in order to populate counters for posts. This
is due the nature how we constructed the service interfaces. In this
change-set we introduce a specialised method to get all counts for a
list of posts at once. After we verify that this does impact the
response times significantly we can mirror the same idea in the cache
layer.